removed isCilium flag

Update run_annotator.sh

Fixed to address review comment

Update ANNOTATION_GENERATOR.md

Updated readme to point to run_annotator.sh

removed run1.sh file


clean up old files
