+++ run_annotator.sh
@@ -2,7 +2,7 @@
 python3 src/annotator.py -o op/katran/txl_katran -s projects/katran/original_source/ -c projects/katran/commented_katran -t op/katran/katran.function_file_list.json -u op/katran/katran.struct_file_list.json -p katran -d projects/katran/boston_katran_comments.json
 
 #cilium
-python3 src/annotator.py -o op/cilium/txl_cilium -s projects/cilium/original_source/ -c op/cilium/commented_cilium -t op/cilium/cilium.function_file_list.json -u op/cilium/cilium.struct_file_list.json --isCilium -p cilium -d projects/cilium/boston_cilium_comments.json
+python3 src/annotator.py -o op/cilium/txl_cilium -s projects/cilium/original_source/ -c op/cilium/commented_cilium -t op/cilium/cilium.function_file_list.json -u op/cilium/cilium.struct_file_list.json -p cilium -d projects/cilium/boston_cilium_comments.json
 
 #bcc
 python3 src/annotator.py -o op/bcc/txl_bcc -s projects/bcc/original_source/ -c op/bcc/commented_bcc -t op/bcc/bcc.function_file_list.json -u op/bcc/bcc.struct_file_list.json  -p bcc
