+++ src/annotator.py
@@ -8,7 +8,6 @@ import subprocess
 import glob
 import shutil
 import code_commentor as cmt
-import summarizer as sm
 import argparse
 import json
 from collections import defaultdict
@@ -135,7 +134,7 @@ def make_cscope_db(db_name,code_dir, cscope_files,cscope_out,tage_folder):
     run_cmd("cqmakedb -s "+ db_name+ " -c "+cscope_out+" -t "+tags_folder+" -p")
 
 # parses output from c-extract-function.txl
-def parseTXLFunctionOutputFile(inputFile, func_file_def_dict, isCilium, helperdict):
+def parseTXLFunctionOutputFile(inputFile, func_file_def_dict, helperdict):
     iFile = open(inputFile,'r')
     lineCt = 1
     srcSeen=False
@@ -180,7 +179,7 @@ def parseTXLFunctionOutputFile(inputFile, func_file_def_dict, isCilium, helperdi
             func_file_def_dict[key].append(fn_def)
     return func_file_def_dict
 
-def create_txl_annotation(cscope_file, opdir,func_file_def_dict, map_file_def_dict, isCilium, helperdict):
+def create_txl_annotation(cscope_file, opdir,func_file_def_dict, map_file_def_dict, helperdict):
     print("Read cscope files and generate function annotation ...")
     txl_dict_func_file = {}
     code_f = open(cscope_file,'r')
@@ -200,7 +199,7 @@ def create_txl_annotation(cscope_file, opdir,func_file_def_dict, map_file_def_di
         #print("File to annotate - ",full_line,"output in",opfile_function_annotate,opfile_struct_annotate)
         op = run_cmd("txl -o "+ opfile_function_annotate+" "+ full_line +" asset/txl/c-extract-functions.txl")
         op = run_cmd("txl -o "+ opfile_struct_annotate +" "+  full_line +" asset/txl/c-extract-struct.txl")
-        func_file_def_dict = parseTXLFunctionOutputFile(opfile_function_annotate, func_file_def_dict, isCilium, helperdict)
+        func_file_def_dict = parseTXLFunctionOutputFile(opfile_function_annotate, func_file_def_dict, helperdict)
         map_file_def_dict = parseTXLStructOutputFile(opfile_struct_annotate, map_file_def_dict)
         txl_dict_func_file[full_line] = opfile_function_annotate
     return func_file_def_dict, txl_dict_func_file, map_file_def_dict
@@ -209,14 +208,9 @@ def create_cqmakedb(db_file, cscope_file, tags_folder):
     run_cmd("cqmakedb -s "+db_file+" -c "+cscope_file+" -t "+tags_folder+" -p")
     return
 
-def create_code_comments(txl_dict, helperdict, opdir, isCilium, human_comments_file, db_file):
-    if(isCilium == False):
-        map_update_fn = ["bpf_sock_map_update", "bpf_map_delete_elem", "bpf_map_update_elem","bpf_map_pop_elem", "bpf_map_push_elem"]
-        map_read_fn = ["bpf_map_peek_elem", "bpf_map_lookup_elem", "bpf_map_pop_elem"]
-    else:
-        map_update_fn = ["sock_map_update", "map_delete_elem", "map_update_elem","map_pop_elem", "map_push_elem"]
-        map_read_fn = ["map_peek_elem", "map_lookup_elem", "map_pop_elem"]
-
+def create_code_comments(txl_dict, helperdict, opdir, human_comments_file, db_file):
+    map_update_fn = ["bpf_sock_map_update", "bpf_map_delete_elem", "bpf_map_update_elem","bpf_map_pop_elem", "bpf_map_push_elem", "sock_map_update", "map_delete_elem", "map_update_elem","map_pop_elem", "map_push_elem"]
+    map_read_fn = ["bpf_map_peek_elem", "bpf_map_lookup_elem", "bpf_map_pop_elem", "map_peek_elem", "map_lookup_elem", "map_pop_elem"]
     funcCapDict = dict()
     for srcFile,txlFile in txl_dict.items():
         opFile = opdir+'/'+os.path.basename(srcFile)
@@ -281,8 +275,6 @@ if __name__ == "__main__":
             help='JSON with information regarding functions present. output of foundation_maker.py')
     my_parser.add_argument('-u','--txl_struct_list',action='store',required=False,
             help='JSON with information regarding structures present. output of foundation_maker.py')
-    my_parser.add_argument('--isCilium', action='store_true',required=False,
-            help='whether repository is cilium')
     my_parser.add_argument('-d','--human_comments_file', action='store',required=False,
             help='JSON with information containing human comments ')
 
@@ -293,9 +285,6 @@ if __name__ == "__main__":
 
     dir_list = []
     
-    isCilium=False
-    if(args.isCilium is True):
-        isCilium = True
     src_dir = args.src_dir
     if (os.access(src_dir, os.R_OK) is not True):
         print("Cannot read source folder: "+src_dir+" Exiting...")
@@ -346,7 +335,7 @@ if __name__ == "__main__":
     tags_folder = "tags"
     #bpf_helper_file = "asset/helper_hookpoint_map.json"
     bpf_helper_file = "asset/bpf_helpers_desc_mod.json"
-    helperdict = sm.load_bpf_helper_map(bpf_helper_file)  
+    helperdict = cmt.load_bpf_helper_map(bpf_helper_file)  
     intermediate_f_list = []
     intermediate_f_list.append(cscope_out)
     intermediate_f_list.append(tags_folder)
@@ -357,14 +346,14 @@ if __name__ == "__main__":
 
     txl_dict_struct = defaultdict(list)
     txl_dict_func = defaultdict(list)
-    txl_dict_func, txl_func_file, txl_dict_struct = create_txl_annotation(cscope_files, txl_op_dir, txl_dict_func, txl_dict_struct, isCilium, helperdict)
+    txl_dict_func, txl_func_file, txl_dict_struct = create_txl_annotation(cscope_files, txl_op_dir, txl_dict_func, txl_dict_struct, helperdict)
     funcCapDict = dict()
     if (cmt_op_dir is not None):
         comments_db_file = cmt_op_dir+"/"+ db_file +"_comments.db"
         comments_db = TinyDB(comments_db_file)
         if(args.bpfHelperFile is not None):
             bpf_helper_file = args.bpfHelperFile
-        funcCapDict = create_code_comments(txl_func_file, helperdict, cmt_op_dir, isCilium, human_comments_file, db_file)
+        funcCapDict = create_code_comments(txl_func_file, helperdict, cmt_op_dir, human_comments_file, db_file)
         funcCapDict = remove_txl_missed_fn(funcCapDict)
         add_root_and_level_info(funcCapDict)
         for fn in funcCapDict.keys():
